move template functionality to separate package
