+++ go.mod
@@ -39,7 +39,7 @@ require (
 	github.com/olekukonko/tablewriter v0.0.5
 	github.com/packethost/packngo v0.29.0
 	github.com/pkg/browser v0.0.0-20210911075715-681adbf594b8
-	github.com/pluralsh/gqlclient v1.3.4
+	github.com/pluralsh/gqlclient v1.3.5
 	github.com/pluralsh/plural-operator v0.5.3
 	github.com/pluralsh/polly v0.0.6
 	github.com/rodaine/hclencoder v0.0.1
